chore: move button to its own file
